FIX. Resolve board controller test
Resovlved the getBoardTest() mock issue. Issue consisted
of trying to fetch and item inside an JSON array and
also parsing the time of the item. Both issues were addressed
although futher fixes have to occur on the way we extarct an item
from the JSON array.

Also this means there were some changes to the Board class
that have to be addressed for consistency and to pass the now
failing tests for BoardTest.java.